### 2024-03-19 #6341 @plaplant
@sarahstangl-lanl by the way, for the paper rendering error, I was able to fix this in a local build by making the figure reference in the paper self-consistent with the figure label. I am not sure if that's what's causing the error in the official rendering, but that might be something worth trying.
