### 2024-04-16 #6637 @plaplant
@guadabsb15 @steven-murray thank you again for agreeing to review this submission! When you get a chance, please comment in this thread:

Then you can go through the checklist items, and make issues on [the upstream repo](https://github.com/lanl/nuDustC) for any changes you recommend. Let me know if you have any questions!

@sarahstangl-lanl it seems like there are still some issues with rendering the PDF, which I think are related to the [figure referencing issue I mentioned in the pre-review thread](https://github.com/openjournals/joss-reviews/issues/6341#issuecomment-2005725596). When you get a chance, please try to update the paper so that the PDF renders correctly, as reviewing the paper is something the reviewers are required to do.

Also, some of the DOIs in the paper references are missing or do not resolve appropriately. See the comment [here](https://github.com/openjournals/joss-reviews/issues/6637#issuecomment-2059428310) for a list of references with issues. If you have questions about how to resolve these issues, please let me know.
