### 2024-04-16 #6341 @sarahstangl-lanl
Sorry about that. I don't know what was wrong with it or how to fix it. So, I just took it out and just said the 'figure'. I'm used to writing in latex.
