### 2024-05-07 lanl/nuDustC#2 @steven-murray
Installation Documentation
Hi! This issue is a part of my JOSS review [here](https://github.com/openjournals/joss-reviews/issues/6637).

Thank you for providing installation instructions, however at least to me they were not super clear. I installed Boost with conda as suggested, however I was expecting the other dependencies to be installed automatically via the `cmake` (this is partly my own unfamiliarity with cmake, but I expect other users may have this same problem). I found that I needed to install `sundials` manually, which turned out to be a bit of a hassle (the link to their own installation docs on their readme is a broken link).

For this repo, it would be useful to at least link to the installation docs for sundials when you mention it as a dependency, and also to mention that it must be manually installed when describing the installation instructions. If possible, I'd also give a short default description of how to install sundials, referring to their documentation for more involved use-cases.

As a further note: I find the layout of the sections on the readme a bit confusing. There is a sub-heading "To Build & Run" and then further down a higher-level heading "Building nuDustC++" which is confusing. I think a high-level heading like "Installation" followed by sub-headings "Dependencies" then "Building nuDustC++" would be useful.
