### 2024-05-07 lanl/nuDustC#3 @steven-murray
Test Run going forever?
Hi! This issue is a part of my JOSS review [here](https://github.com/openjournals/joss-reviews/issues/6637).

After successfully (?) building `nuDustC++`, I tried the instructions: `$> ./nudust++ -c data/inputs/test_config.ini`. The code started to run and I saw a nice ASCII title text. But when the code got to the "Starting Nucleation" part, it sat there for several hours, before I manually cancelled it.

It's quite possible that it should just take that long to run, but there is no indication on the readme for what to expect, and there are no additional updates output by the code to let me know it's still running properly. It would be great if these could be added (and if it really does take this long, is there a smaller example that could be run to test that the code works?).
