### 2024-05-07 lanl/nuDustC#3 @guadabsb15
Hi! I also had a few thoughts on this topic.  I was able to run serially, and it took ~20 minutes for me, but I also think it would be nice to have a shorter example to test the installation. I got the following output in the `log.txt` file

```
2024-04-23 08:27:55.379 INFO  [42298] [main@122] nuDust has started
2024-04-23 08:27:55.381 INFO  [42298] [nuDust::load_network@79] loaded network file
2024-04-23 08:27:55.382 INFO  [42298] [nuDust::load_initial_abundances@312] loaded abundance file. loaded 87 abundances.
2024-04-23 08:27:55.386 INFO  [42298] [nuDust::load_environment_data@476] loaded environment file
2024-04-23 08:27:55.386 INFO  [42298] [nuDust::generate_sol_vector@568] generated solution vector
2024-04-23 08:27:55.386 INFO  [42298] [nuDust::load_outputFL_names@585] data and restart file names defined
2024-04-23 08:27:55.401 INFO  [42298] [nuDust::create_simulation_cells@645] Created 6 cells
2024-04-23 08:27:55.401 INFO  [42298] [nuDust::run@652] Entering main integration loop
2024-04-23 08:27:55.401 DEBUG [42298] [cell::solve@165] Times taken from from environment file.
2024-04-23 08:30:57.831 INFO  [42298] [cell::solve@195] finished integration cell: 1, t_current: 2.60412e+07
2024-04-23 08:30:57.831 INFO  [42298] [cell::solve@230] done cell: 1
2024-04-23 08:30:57.832 DEBUG [42298] [cell::solve@165] Times taken from from environment file.
2024-04-23 08:34:01.272 INFO  [42298] [cell::solve@195] finished integration cell: 2, t_current: 2.60412e+07
2024-04-23 08:34:01.272 INFO  [42298] [cell::solve@230] done cell: 2
2024-04-23 08:34:01.272 DEBUG [42298] [cell::solve@165] Times taken from from environment file.
2024-04-23 08:37:05.345 INFO  [42298] [cell::solve@195] finished integration cell: 3, t_current: 2.60412e+07
2024-04-23 08:37:05.345 INFO  [42298] [cell::solve@230] done cell: 3
2024-04-23 08:37:05.345 DEBUG [42298] [cell::solve@165] Times taken from from environment file.
2024-04-23 08:40:08.371 INFO  [42298] [cell::solve@195] finished integration cell: 4, t_current: 2.60412e+07
2024-04-23 08:40:08.371 INFO  [42298] [cell::solve@230] done cell: 4
2024-04-23 08:40:08.372 DEBUG [42298] [cell::solve@165] Times taken from from environment file.
2024-04-23 08:43:11.518 INFO  [42298] [cell::solve@195] finished integration cell: 5, t_current: 2.60412e+07
2024-04-23 08:43:11.518 INFO  [42298] [cell::solve@230] done cell: 5
2024-04-23 08:43:11.518 DEBUG [42298] [cell::solve@165] Times taken from from environment file.
2024-04-23 08:46:15.634 INFO  [42298] [cell::solve@195] finished integration cell: 6, t_current: 2.60412e+07
2024-04-23 08:46:15.634 INFO  [42298] [cell::solve@230] done cell: 6
2024-04-23 08:46:15.634 INFO  [42298] [nuDust::run@682] Leaving main integration loop
```
And the following files in the `output` directory  `B100_test_chm_1.dat  B100_test_chm_2.dat  B100_test_chm_3.dat  B100_test_chm_4.dat  B100_test_chm_5.dat  B100_test_chm_6.dat`

But I'm not sure if those outputs are "correct" or not. It would be nice to have a short example to run and the reference on what to expect in the `log.txt` . Also maybe a script that does some quick read/analysis of the output files with a reference plot or table or quantity, to have a sanity check.
