### 2024-05-14 lanl/nuDustC#2 @guadabsb15
Just adding a couple of thoughts that came to my mind when going through the installation instructions. I agree with @steven-murray comments here, and I think it could be valuable to add in the documentation some extra information. For example, in the case of sundials installation, to enable MPI and OpenMP one has to change `ENABLE_MPI`  and  `ENABLE_OPENMP` to `ON`. I'm not sure if it is necessary to change other options from their default values, but a note about that would be helpful.
In general, it would also be a good idea to mention which version/release of the dependencies you have used or tested. Similarly for compilers and MPI implementations.
