### 2024-05-14 #6637 @sarahstangl-lanl
Is number 3 the "reproducibility" from the checklist? I added a pdf version of the paper: https://github.com/lanl/nuDustC/blob/main/joss_nudustc.pdf

What is the issue with rendering the image in the paper? It comes up for me using markdown.
