### 2024-05-15 #6637 @steven-murray
@sarahstangl-lanl it would be good if the paper could be generated in the JOSS style as well.
