### 2024-05-16 lanl/nuDustC#4 @guadabsb15
Not able to build with MPI and/or OpenMP
Hi! This issue is also related to the JOSS review https://github.com/openjournals/joss-reviews/issues/6637

Thank you for the description provided for installation, although I was able to build nuDustC serially, when trying to build with `NUDUSTC_ENABLE_MPI=ON` , I get the following error :
```
CMake Error at CMakeLists.txt:86 (target_link_libraries):
  Target "nudustc++" links to:

    "MPI::MPI_CXX"

  but the target was not found.  Possible reasons include:

    * There is a typo in the target name.
    * A find_package call is missing for an IMPORTED target.
    * An ALIAS target is missing.
```
I get an analogous error when using `NUDUSTC_ENABLE_OPENMP=ON` I tried both using cmake 3.29 and cmake 3.23.3, but it looks like the quotes inside the generator  `$<${with_mpi}:"MPI::MPI_CXX">` in `CMakeLists.txt` might be causing an issue. CMake seems to find the path to the MPI relevant paths, so I'm not sure that is the problem.
