### 2024-05-22 #6637 @plaplant
@sarahstangl-lanl it looks like there is a `paper.md` file both in the top-level of the directory, as well as inside of the `JOSS Paper` directory. In addition to duplicating the paper content (which we should probably avoid), it seems like the space in the directory name is what is causing the paper generation to fail. Would you be able to reorganize the repo so that there is only a single copy of the `paper.md` file? And if it's in a directory, please rename the directory so that there are no spaces in the full path to the `paper.md` file. Thanks!
