+++ data/inputs/config.ini
@@ -3,8 +3,6 @@ ode_dt_min = 1.0E-12
 ode_dt_max = 1
 ode_abs_err = 1.0E-5
 ode_rel_err = 1.0E-5
-io_disk_n_steps = 10000
-io_screen_n_steps = 10000
 network_file = data/chemistry/test_chm.chm
 abundance_file = data/abundances/test_abund.dat
 environment_file = data/trajectories/test_traj.dat
