### 2024-05-29 dd146d9
Delete inputs directory
