+++ README.md
@@ -36,7 +36,9 @@ On MacOS using HomeBrew, run:
 brew install sundials
 ```
 
-Installation instructions for sundials can be found here: [sundials](https://computing.llnl.gov/projects/sundials)
+Installation instructions for sundials can be found at: 
+
+https://computing.llnl.gov/projects/sundials
 
 # Building nuDustC++
 In order to build nuDustC++, go to the head of the git repository (nudustc/) and run:
