+++ README.md
@@ -7,7 +7,7 @@ nuDustC++ is used to calculate the dust neclation and destruction in gaseous sys
 
 ***All Units are in CGS***
 
-### To Build & Run
+### Requirements
 **Required:** OpenMP, MPI, Boost, SunDials, Plog.
 This build uses cmake. 
 
