+++ README.md
@@ -41,7 +41,7 @@ Installation instructions for sundials can be found at:
 https://computing.llnl.gov/projects/sundials
 
 # Building nuDustC++
-In order to build nuDustC++, go to the head of the git repository (nudustc/) and run:
+In order to build nuDustC++, go to the head of the git repository (nuDustC/) and run:
 
 ```
 $> mkdir build;
