### 2024-05-29 lanl/nuDustC#2 @sarahstangl-lanl
I added and cleaned up the documentation's installation  section and added information on enabling OpenMP and MPI. This hopefully is clearer now.
