### 2024-05-29 2f695e8
Delete JOSS Paper directory
