+++ JOSS Paper/paper.md
@@ -1,78 +1,0 @@
----
-title: '<span style="font-family:qcr;">nudustc++</span>: C++ Code for Modeling Dust Nucleation and Destruction in Gaseous Sysytems'
-tags:
-  - astronomy
-  - nucleation
-  - C++
-  - Supernovae
-  - astrophysics
-  - dust
-  - dust grains
-  - grain physics
-  - physics
-authors:
-  - name: Sarah M. Stangl
-    orcid: 0000-0001-5570-6666
-    equal-contrib: true
-    affiliation: "1, 2,3" 
-  - name: Christopher M. Mauney
-    orcid: 0000-0002-7827-2247
-    equal-contrib: true 
-    affiliation: "1,2,4"
-affiliations:
- - name: Theoretical Design Division, Los Alamos National Laboratory, Los Alamos, NM, 87545, USA
-   index: 1
- - name: Center for Theoretical Astrophysics, Los Alamos National Laboratory, Los Alamos, NM, 87545, USA
-   index: 2
- - name: University of Oklahoma, Norman, OK, 73019, USA
-   index: 3
- - name: High Performance Computing Division, Los Alamos National Laboratory, Los Alamos, NM, 87545, USA
-   index: 4
-date: 8 January 2024
-bibliography: paper.bib
----
-
-# Summary
-
-We introduce <span style="font-family:qcr;">nudustc++</span>, a **nu**cleating **dust** code in **C++** modeling dust grain formation, growth, and erosion in gaseous systems. <span style="font-family:qcr;">nudustc++</span> is a highly parallizable set of code and tools for solving a system of nonlinear ordinary differential equations describing dust nucleation, growth, and erosion for user-specified grain species. It leverages OpenMP and MPI to optimize threading and distribution on available CPUS.
-
-# Statement of need
-
-Understanding interstellar dust is crucial for astronomical observations [@Draine2003], offering key insights into stellar processes. These grains absorb electromagnetic radiation, re-emitting it at longer wavelengths, leading to extinction and a spectral shift towards redder wavelengths. The size and composition of dust introduces variability in opacities and distortion of incident light, resulting in molecular lines and altering the resulting data. Dust formed in asymptotic giant branch (AGB) stars, on pre-existing grains in the interstellar medium (ISM), and within the expanding, cooling ejecta of core collapse supernovae explosions (CCSNe); these grains preserve important information about the nucleosynthetic processes within their host environment, locking up their unique isotopic signatures. However, despite their importance, the quantity, composition, and size distribution of dust formed in supernovae and deposited into the ISM remain poorly constrained. 
-
-Models of formation, growth, and weathering of dust are necessary to understand the origin and characteristics of dust in the interstellar medium, shedding light on where and what dust are possible in these environments. Specifically, modelling the formation and survivalability of dust in CCSNe produce a population of dust grains that can be compared with observations, allowing verification of our current understanding of physical models: ISM dust origin, chemical networks, nucleation models, hydrodynamics, supernovae engines, progenitor structure, erosion physics, and stellar compositions. 
-
[…]
