+++ JOSS Paper/paper.bib
@@ -1,178 +1,0 @@
-@MISC{astrochem,
-       author = {{Maret}, S{\'e}bastien and {Bergin}, Edwin A.},
-        title = "{Astrochem: Abundances of chemical species in the interstellar medium}",
-     keywords = {Software},
- howpublished = {Astrophysics Source Code Library, record ascl:1507.010},
-         year = 2015,
-        month = jul,
-          eid = {ascl:1507.010},
-        pages = {ascl:1507.010},
-archivePrefix = {ascl},
-       eprint = {1507.010},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2015ascl.soft07010M},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{Draine2003,
-       author = {{Draine}, B.~T.},
-        title = "{Interstellar Dust Grains}",
-      journal = {\araa},
-     keywords = {Astrophysics},
-         year = 2003,
-        month = jan,
-       volume = {41},
-        pages = {241-289},
-          doi = {10.1146/annurev.astro.41.011802.094840},
-archivePrefix = {arXiv},
-       eprint = {astro-ph/0304489},
- primaryClass = {astro-ph},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2003ARA&A..41..241D},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@article{kozasa1987grain,
-  title={Grain formation through nucleation process in astrophysical environments. II: Nucleation and grain growth accompanied by chemical reaction},
-  author={Kozasa, Takahashi and Hasegawa, Hiroichi},
-  journal={Progress of Theoretical Physics},
-  volume={77},
-  number={6},
-  pages={1402--1410},
-  year={1987},
-  publisher={Oxford University Press}
-}
-
-@ARTICLE{Bianchi2007,
-       author = {{Bianchi}, Simone and {Schneider}, Raffaella},
-        title = "{Dust formation and survival in supernova ejecta}",
-      journal = {\mnras},
-     keywords = {shock waves, supernovae: individual: Cassiopeia A, dust, extinction, supernova remnants, Astrophysics},
-         year = 2007,
-        month = jul,
-       volume = {378},
-       number = {3},
-        pages = {973-982},
-          doi = {10.1111/j.1365-2966.2007.11829.x},
-archivePrefix = {arXiv},
-       eprint = {0704.0586},
- primaryClass = {astro-ph},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2007MNRAS.378..973B},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{Todini2001,
-       author = {{Todini}, Paolo and {Ferrara}, Andrea},
-        title = "{Dust formation in primordial Type II supernovae}",
-      journal = {\mnras},
-     keywords = {STARS: FORMATION, SUPERNOVAE: GENERAL, SUPERNOVAE: INDIVIDUAL: SN1987A, DUST, EXTINCTION, SUPERNOVA REMNANTS, Astrophysics},
-         year = 2001,
-        month = aug,
-       volume = {325},
-       number = {2},
-        pages = {726-736},
-          doi = {10.1046/j.1365-8711.2001.04486.x},
-archivePrefix = {arXiv},
-       eprint = {astro-ph/0009176},
- primaryClass = {astro-ph},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2001MNRAS.325..726T},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@INPROCEEDINGS{paquette2023,
[…]
