### 2024-05-29 #6637 @sarahstangl-lanl
I've removed the directory. The paper is only at the main level.
