### 2024-05-30 #6637 @plaplant
@sarahstangl-lanl thanks for reorganizing the repo! The paper generation succeeded 🎉

@guadabsb15 @steven-murray thanks very much for starting your reviews and providing feedback! The paper is now available in [this comment](https://github.com/openjournals/joss-reviews/issues/6637#issuecomment-2138450854).
