### 2024-06-18 lanl/nuDustC#4 @sarahstangl-lanl
I'm unable to recreate this error. It looks like there is an issue with cmake finding MPI. You might need to specify the path to the MPI build and ensure the correct MPI is called in CMake is installed.
