### 2024-07-04 lanl/nuDustC#4 @steven-murray
I also am having trouble installing on Ubuntu. My error is that it just can't find `mpi.h`, but I have installed `MPICH` with `apt`. I can't find where to specify the include directory for MPI. This error occurs even if I set

```
option(NUDUSTC_ENABLE_MPI OFF "Use MPI for cell/particle parallelization")
```

in the `CMakeLists.txt` (it is `ON` by default, despite what the readme says).
