### 2024-07-04 lanl/nuDustC#4 @steven-murray
I tried again on Manjaro linux (essentialy Arch), and got a different MPI-based error:

```
[ 12%] Building CXX object CMakeFiles/nudustc++.dir/src/cell.cpp.o
[ 25%] Building CXX object CMakeFiles/nudustc++.dir/src/cellobserver.cpp.o
[ 37%] Building CXX object CMakeFiles/nudustc++.dir/src/configuration.cpp.o
/home/sgm/work/joss-reviews/nudustc/src/configuration.cpp: In constructor ‘configuration::configuration()’:
/home/sgm/work/joss-reviews/nudustc/src/configuration.cpp:39:102: warning: overflow in conversion from ‘float’ to ‘int’ changes value from ‘+QNaNf’ to ‘0’ [-Woverflow]
   39 |     desc.add_options() ( "number_of_size_bins", options::value<int> ( &bin_number )->default_value ( NAN ), "bin number" );
      |                                                                                                      ^~~
[ 50%] Building CXX object CMakeFiles/nudustc++.dir/src/main.cpp.o
[ 62%] Building CXX object CMakeFiles/nudustc++.dir/src/network.cpp.o
[ 75%] Building CXX object CMakeFiles/nudustc++.dir/src/nudust.cpp.o
[ 87%] Building CXX object CMakeFiles/nudustc++.dir/src/reaction.cpp.o
[100%] Linking CXX executable nudustc++
/usr/sbin/ld: CMakeFiles/nudustc++.dir/src/main.cpp.o: warning: relocation against `ompi_mpi_comm_world' in read-only section `.text'
/usr/sbin/ld: CMakeFiles/nudustc++.dir/src/main.cpp.o: in function `main':
main.cpp:(.text+0x7e): undefined reference to `MPI_Init'
/usr/sbin/ld: main.cpp:(.text+0x97): undefined reference to `MPI_Init'
/usr/sbin/ld: main.cpp:(.text+0x117): undefined reference to `ompi_mpi_comm_world'
/usr/sbin/ld: main.cpp:(.text+0x11f): undefined reference to `MPI_Comm_rank'
/usr/sbin/ld: main.cpp:(.text+0x136): undefined reference to `ompi_mpi_comm_world'
/usr/sbin/ld: main.cpp:(.text+0x13e): undefined reference to `MPI_Comm_size'
/usr/sbin/ld: warning: creating DT_TEXTREL in a PIE
collect2: error: ld returned 1 exit status
make[2]: *** [CMakeFiles/nudustc++.dir/build.make:197: nudustc++] Error 1
make[1]: *** [CMakeFiles/Makefile2:100: CMakeFiles/nudustc++.dir/all] Error 2
make: *** [Makefile:91: all] Error 2
```

Again, this happens whether or not I switch on MPI in the `CMakeLists.txt`.

This is with `gcc 14`  and openmpi 5
