### 2024-07-09 fbb21ca
removed extra checks
