+++ src/main.cpp
@@ -57,30 +57,11 @@ and chemistry.
 
 int main(int argc, char *argv[]) {
   int rank = 0, size = 1;
-
-  std::cout << "Before MPI_Init" << std::endl;
 //#ifdef NUDUSTC_ENABLE_MPI
   MPI_Init(&argc, &argv);
 
-  int mpi_initialized = MPI_Init(&argc, &argv);
-  if (mpi_initialized != MPI_SUCCESS) {
-      std::cerr << "MPI initialization failed!" << std::endl;
-      return -1;
-  }
-  std::cout << "After MPI_Init" << std::endl;
-
-  int rank_result = MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-  int size_result = MPI_Comm_size(MPI_COMM_WORLD, &size);
-
-  if (rank_result != MPI_SUCCESS) {
-      std::cerr << "MPI_Comm_rank failed!" << std::endl;
-  }
-  if (size_result != MPI_SUCCESS) {
-      std::cerr << "MPI_Comm_size failed!" << std::endl;
-  }
-
-  std::cout << "rank: " << rank << std::endl;
-  std::cout << "size: " << size << std::endl;
+  MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+  MPI_Comm_size(MPI_COMM_WORLD, &size);
 //#endif
   namespace po = boost::program_options;
 
