### 2024-07-09 be3e650
removed mpi_root specification
