### 2024-07-09 c5a0215
edited MPI instructions
