### 2024-07-10 lanl/nuDustC#4 @guadabsb15
With recent changes in the repository, I also get the errors @steven-murray mentions. I'm using ubuntu and gcc14, and tried with both openmpi 5 and 4.1
