### 2024-07-10 lanl/nuDustC#4 @sarahstangl-lanl
Can you try again? CMake is finding my mpi. I added lines to help cmake find the users mpi in the CMakeLists.txt. Are you using openmpi or mpich?

Does adding the mpi build location to you path help?

export PATH="/pathTo/open-mpi/bin:$PATH"
export CPATH="/pathTo/open-mpi/include:$CPATH"
export LIBRARY_PATH="/pathTo/open-mpi/lib:$LIBRARY_PATH"
export LD_LIBRARY_PATH="/pathTo/open-mpi/lib:$LD_LIBRARY_PATH"
