### 2024-07-11 lanl/nuDustC#4 @guadabsb15
I just tried again "cmake -DNUDUSTC_ENABLE_MPI=ON" to make sure that MPI is enabled. The output I get is
```
MPI Found
-- Configuring done
CMake Error at CMakeLists.txt:91 (target_link_libraries):
  Target "nudustc++" links to:

    "MPI::MPI_CXX"

  but the target was not found.  Possible reasons include:

    * There is a typo in the target name.
    * A find_package call is missing for an IMPORTED target.
    * An ALIAS target is missing.
```
So MPI is found, as shown by the recent message you added for that purpose. Also, looking at the CMakeCache.txt in the build directory I can see `MPI_mpi_LIBRARY:FILEPATH=/home/mgbarrioss/miniforge3/envs/joss_nuDustc_3/lib/libmpi.so`

However the original issue message remains in the output.

As I said originally I suspect the problem is with the quotes around `MPI::MPI_CXX` in line 99 of CMakeLists.txt. If I remove those quotes, the build process finishes.
