### 2024-07-11 lanl/nuDustC#6 @guadabsb15
Code no longer builds serially 
This comment is also part of the review https://github.com/openjournals/joss-reviews/issues/6637

With the changes in commit `54875fa`, the preprocessor directives for `#ifdef NUDUSTC_ENABLE_MPI` were commented out in src/main.cpp at lines 21 and 23, 60 and 65. Therefore, this code sections are being compiled even when building serially.

If these lines are uncommented, the code can be built again by running `cmake -DNUDUSTC_ENABLE_MPI=OFF ..` and `make`
