### 2024-07-19 #6637 @plaplant
@sarahstangl-lanl 👋 just checking in on the status of this review. It sounds like the reviewers are having issues building and installing the code (from the above issues). Do you know if they have been resolved? It sounds like these issues might be blocking further review of the code.
