### 2024-07-21 #6637 @steven-murray
@plaplant yes there is some discussion about installation issues over on the repo -- I will try the installation process again in the coming week.
