### 2024-10-10 #6637 @plaplant
@sarahstangl-lanl as we have not heard from you in several months, I am going to go ahead and withdraw this submission from consideration for publication. You are able to re-submit this for publication in the future if you decide you would like to publish in JOSS.

@steven-murray @guadabsb15 thank you very much for the time and effort invested in reviewing this submission. We appreciate your help with this review! Because this submission will be withdrawn, no further action is needed.
